+++ tests/test_tls_integration.c
@@ -38,44 +38,44 @@ int test_tls_integration(void) {
   const char *msgs[2];
   int i;
 
-  res = c_rest_tls_init();
+  res = (int)c_rest_tls_init();
   failed += (res != C_REST_OK);
 
   /* Test malloc failures */
   g_crf_malloc_hook = fail_malloc_n;
   for (i = 0; i < 5; i++) {
     g_malloc_fail_after = i;
-    res = c_rest_tls_context_init(&tls_ctx);
+    res = (int)c_rest_tls_context_init(&tls_ctx);
     if (res == C_REST_OK) {
       c_rest_tls_context_destroy(tls_ctx);
     }
   }
   g_crf_malloc_hook = NULL;
 
-  res = c_rest_tls_context_init(&tls_ctx);
+  res = (int)c_rest_tls_context_init(&tls_ctx);
   failed += (res != C_REST_OK);
 
   if (res == C_REST_OK) {
-    res = c_rest_tls_load_cert(tls_ctx, "tests/certs/server.crt");
-    res = c_rest_tls_load_key(tls_ctx, "tests/certs/server.key");
+    res = (int)c_rest_tls_load_cert(tls_ctx, "tests/certs/server.crt");
+    res = (int)c_rest_tls_load_key(tls_ctx, "tests/certs/server.key");
 
-    res = c_rest_init(C_REST_MODALITY_SYNC, &ctx);
+    res = (int)c_rest_init(C_REST_MODALITY_SYNC, &ctx);
     failed += (res != C_REST_OK);
     if (res == C_REST_OK) {
       ctx->tls_ctx = tls_ctx;
-      res = c_rest_destroy(ctx);
+      res = (int)c_rest_destroy(ctx);
       failed += (res != C_REST_OK);
     }
 
-    res = c_rest_init(C_REST_MODALITY_ASYNC, &ctx);
+    res = (int)c_rest_init(C_REST_MODALITY_ASYNC, &ctx);
     failed += (res != C_REST_OK);
     if (res == C_REST_OK) {
       ctx->tls_ctx = tls_ctx;
-      res = c_rest_destroy(ctx);
+      res = (int)c_rest_destroy(ctx);
       failed += (res != C_REST_OK);
     }
 
-    res = c_rest_tls_context_destroy(tls_ctx);
+    res = (int)c_rest_tls_context_destroy(tls_ctx);
     failed += (res != C_REST_OK);
   }
